requirements.txt
fastapi
uvicorn
google-cloud-bigquery
python-dotenv
pydantic